Add scaling parameter, make number cores an integer

+++ src/ParameterStudyVariants.py
@@ -252,7 +252,7 @@ class ChangeNumberOfSubdomains(Variant):
         self.number_cores = self.value[0]
         if isinstance(self.number_cores, str):
             if self.number_cores == "fullNode":
-                self.number_cores = multiprocessing.cpu_count() / 2
+                self.number_cores = int(multiprocessing.cpu_count() / 2)
         name = str(self.number_cores)
         super().__init__(
             root_dir,
